Fix merge conflict from develop
